refactor: Modularize setup script by extracting shared utilities and service-specific logic into dedicated modules.
